test: separate out unit tests and system tests
